requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
python-dotenv
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
